Code cleanup

+++ libmirai/private/libmirai/Impl/MessageClientImpl.cpp
@@ -1,15 +1,15 @@
 // Copyright (C) 2022 Numendacil and contributors
-// 
+//
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU Affero General Public License as
 // published by the Free Software Foundation, either version 3 of the
 // License, or (at your option) any later version.
-// 
+//
 // This program is distributed in the hope that it will be useful,
 // but WITHOUT ANY WARRANTY; without even the implied warranty of
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU Affero General Public License for more details.
-// 
+//
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
@@ -21,7 +21,7 @@
 namespace Mirai::Details
 {
 
-bool MessageClientImpl::_init = false;
+bool MessageClientImpl::_init = false;	// NOLINT(cppcoreguidelines-avoid-non-const-global-variables)
 
 MessageClientImpl::MessageClientImpl(std::chrono::seconds HeartbeatInterval, std::chrono::seconds HandshakeTimeout,
                                      bool EnablePong, bool EnableDeflate, bool AutoReconnect,
@@ -33,8 +33,8 @@ MessageClientImpl::MessageClientImpl(std::chrono::seconds HeartbeatInterval, std
 		ix::initNetSystem();
 		MessageClientImpl::_init = true;
 	}
-	this->_client.setPingInterval(HeartbeatInterval.count());
-	this->_client.setHandshakeTimeout(HandshakeTimeout.count());
+	this->_client.setPingInterval(static_cast<int>(HeartbeatInterval.count()));
+	this->_client.setHandshakeTimeout(static_cast<int>(HandshakeTimeout.count()));
 	if (!EnablePong) this->_client.disablePong();
 	if (EnableDeflate) this->_client.enablePerMessageDeflate();
 	if (AutoReconnect)
